Enhance password reset
